test: move tests into main
