+++ src/test/main_test.ts
@@ -3,8 +3,11 @@ import {describe, test, expect} from 'vitest';
 import os from 'node:os';
 import fs from 'node:fs';
 import path from 'node:path';
+import {spawnSync} from 'node:child_process';
 
 const isWindows = os.platform() === 'win32';
+const fixturesDir = path.join(import.meta.dirname, '../../test/fixtures');
+const distDir = path.join(import.meta.dirname, '../../dist');
 
 const variants = [
   {name: 'async', x, isAsync: true},
@@ -360,6 +363,82 @@ if (!isWindows) {
         expect(result.exitCode).toBe(0);
         expect(result.stdout).toBe('local\n');
       } finally {
+        fs.rmSync(dir, {recursive: true, force: true});
+      }
+    });
+
+    test('resolves when grandchild holds piped stdout open', async () => {
+      const dir = fs.mkdtempSync(
+        path.join(os.tmpdir(), 'tinyexec-grandchild-')
+      );
+      const runnerScript = path.join(dir, 'runner.mjs');
+      const distPath = JSON.stringify(path.join(distDir, 'main.mjs'));
+      const fixturePath = JSON.stringify(
+        path.join(fixturesDir, 'grandchild.mjs')
+      );
+
+      fs.writeFileSync(
+        runnerScript,
+        `import { x } from ${distPath}
+  const result = await x('node', [${fixturePath}])
+  process.stdout.write(JSON.stringify({ stdout: result.stdout, exitCode: result.exitCode }))
+  `
+      );
+
+      try {
+        const proc = spawnSync('node', [runnerScript], {
+          timeout: 10000,
+          encoding: 'utf8',
+          killSignal: 'SIGKILL',
+          stdio: ['pipe', 'pipe', 'pipe']
+        });
+
+        expect(proc.signal).not.toBe('SIGKILL');
+        expect(proc.status).toBe(0);
+        const parsed = JSON.parse(proc.stdout.trim());
+        expect(parsed.exitCode).toBe(0);
+        expect(parsed.stdout).toBe('output\n');
+      } finally {
+        spawnSync('pkill', ['-f', 'tinyexec-test-grandchild']);
+        fs.rmSync(dir, {recursive: true, force: true});
+      }
+    });
+
+    test('iterator completes when grandchild holds piped stdout open', async () => {
+      const dir = fs.mkdtempSync(
+        path.join(os.tmpdir(), 'tinyexec-grandchild-')
+      );
+      const runnerScript = path.join(dir, 'runner.mjs');
+      const distPath = JSON.stringify(path.join(distDir, 'main.mjs'));
+      const fixturePath = JSON.stringify(
+        path.join(fixturesDir, 'grandchild_multiline.mjs')
+      );
+
+      fs.writeFileSync(
+        runnerScript,
+        `import { x } from ${distPath}
+  const lines = []
+  for await (const line of x('node', [${fixturePath}])) {
+    lines.push(line)
+  }
+  process.stdout.write(JSON.stringify(lines))
+  `
+      );
+
+      try {
+        const proc = spawnSync('node', [runnerScript], {
+          timeout: 10000,
+          encoding: 'utf8',
+          killSignal: 'SIGKILL',
+          stdio: ['pipe', 'pipe', 'pipe']
+        });
+
+        expect(proc.signal).not.toBe('SIGKILL');
+        expect(proc.status).toBe(0);
+        const parsed = JSON.parse(proc.stdout.trim());
+        expect(parsed).toEqual(['line1', 'line2']);
+      } finally {
+        spawnSync('pkill', ['-f', 'tinyexec-test-grandchild']);
         fs.rmSync(dir, {recursive: true, force: true});
       }
     });
